+++ packages/web/src/middleware/render-web.ts
@@ -3,12 +3,12 @@ import _ from 'lodash';
 import buildDebug from 'debug';
 import express from 'express';
 
-import { isHTTPProtocol } from '@verdaccio/utils';
 import { SearchInstance } from '@verdaccio/store';
 import { HTTP_STATUS } from '@verdaccio/commons-api';
 import { loadPlugin } from '@verdaccio/loaders';
 import renderHTML from '../renderHTML';
 import { setSecurityWebHeaders } from './security';
+import { isHTTPProtocol } from '../utils/web-utils';
 
 const debug = buildDebug('verdaccio:web:render');
