+++ packages/web/src/utils/web-utils.ts
@@ -84,3 +84,41 @@ export function parseReadme(packageName: string, readme: string): string | void 
   // logger.error({ packageName }, '@{packageName}: No readme found');
   throw new Error('ERROR: No README data found!');
 }
+
+/**
+ * Check if URI is starting with "http://", "https://" or "//"
+ * @param {string} uri
+ */
+export function isHTTPProtocol(uri: string): boolean {
+  return /^(https?:)?\/\//.test(uri);
+}
+
+export function deleteProperties(propertiesToDelete: string[], objectItem: any): any {
+  _.forEach(propertiesToDelete, (property): any => {
+    delete objectItem[property];
+  });
+
+  return objectItem;
+}
+
+export function addScope(scope: string, packageName: string): string {
+  return `@${scope}/${packageName}`;
+}
+
+export function sortByName(packages: any[], orderAscending: boolean | void = true): string[] {
+  return packages.slice().sort(function (a, b): number {
+    const comparatorNames = a.name.toLowerCase() < b.name.toLowerCase();
+    return orderAscending ? (comparatorNames ? -1 : 1) : comparatorNames ? 1 : -1;
+  });
+}
+
+/**
+ * Detect running protocol (http or https)
+ */
+export function getWebProtocol(headerProtocol: string | void, protocol: string): string {
+  if (typeof headerProtocol === 'string' && headerProtocol !== '') {
+    const commaIndex = headerProtocol.indexOf(',');
+    return commaIndex > 0 ? headerProtocol.substr(0, commaIndex) : headerProtocol;
+  }
+  return protocol;
+}
